Baselining errors caused by missing `doctrine/migrations` dependency (not installed)
`doctrine/migrations` is supported in multiple versions/flavors, and therefore we duck-type our way
through a setup here.

This is sub-optimal, and needs a long-term solution later on, but for now, we roll with this approach,
baselining errors related to unknown classes.

A better approach would really be to split up the package into multiple "checks", each to be installed
individually, but that leads to terrible usability/maintainability too.

Discussion around this is deferred.