Simplified and documented data structures of all checks and their constructors
This mostly replaces many docblock-only union types with PHP 8 union types (now usable),
and removes many `Traversable` inputs where unusable.

Specifically, in order to operate properly, many of these checks do things like
`$this->settings[0]` or `count($this->settings)`, even in happy path scenarios,
leading to invalid execution.

This means that the accepted input type is not `array|Traversable` for many of our checks,
but rather `non-empty-list` or stricter.

While this could potentially mean that this is a BC break, these checks did **NOT**
work, when a `Traversable` was given as input, and the errors change from being an
`Exception` to being a more specific `TypeError`.

Tests verifying impossible types/scenarios have been removed.