Handle PHP8.3 parse-error of ini-files
PHP8.3 now emits a syntax error warning in addition to returning
`false` now.
Previously the error was silently ignored and the returned value was an
empty array.

The `@` silcences the warning. As this is done during the check
whether the ini-file can be parsed or not it should be safe here to not
output the content of the warning.

Signed-off-by: Andreas Heigl <andreas@heigl.org>